# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/preview:
    post:
      operationId: previewOnboarding
      tags: [onboarding]
      summary: Compute a personalized program from quiz answers (anonymous, no auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Personalized program preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingProgram"
  /onboarding:
    post:
      operationId: saveOnboarding
      tags: [onboarding]
      summary: Persist quiz answers to the signed-in user and return their program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Saved user with personalized program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /curriculum/{day}:
    get:
      operationId: getCurriculumDay
      tags: [curriculum]
      summary: Get a single curriculum day with its quiz
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Curriculum day detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurriculumDayDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /curriculum/{day}/audio:
    post:
      operationId: generateDayAudio
      tags: [curriculum]
      summary: Generate (or fetch cached) audio deep-dive for a day
      description: Generates a TTS audio deep-dive once per day and caches it; subsequent calls return the cached audio URL.
      parameters:
        - name: day
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audio ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayAudio"
        "404":
          description: Curriculum day not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /feed:
    get:
      operationId: listFeed
      tags: [community]
      summary: List community feed posts (newest first)
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedPost"
    post:
      operationId: createPost
      tags: [community]
      summary: Share a win or reflection to the community feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
  /feed/{id}:
    delete:
      operationId: deletePost
      tags: [community]
      summary: Delete one of your own posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /feed/{id}/like:
    post:
      operationId: likePost
      tags: [community]
      summary: Like a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: unlikePost
      tags: [community]
      summary: Remove your like from a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Me:
      type: object
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        currentDay:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        streakCount:
          type: integer
        streakFreezes:
          type: integer
        weeklyXp:
          type: integer
        isPublicProfile:
          type: boolean
        hideFromLeaderboard:
          type: boolean
        anonymousMode:
          type: boolean
        emailRemindersEnabled:
          type: boolean
        lastActiveDate:
          type: ["string", "null"]
        theme:
          type: string
        hasPaid:
          type: boolean
        isGrandfathered:
          type: boolean
        hasAccess:
          type: boolean
          description: True if the user can access all 28 days (paid or grandfathered).
        accessibleThroughDay:
          type: integer
          description: Highest day number the user can currently open (1 for free preview, 28 if entitled).
        onboardingCompleted:
          type: boolean
        workContext:
          type: ["string", "null"]
        programTitle:
          type: ["string", "null"]
        programFocus:
          type: ["string", "null"]
        dailyMinutes:
          type: integer
          description: Target minutes the learner wants to spend per day (10, 20, or 40). Drives lesson depth and day-completion requirements.
        field:
          type: ["string", "null"]
          description: The domain the learner wants to apply AI to (e.g. marketing, engineering).
        experienceLevel:
          type: ["string", "null"]
          description: Self-reported AI experience (beginner, some, advanced).
        goal:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - currentDay
        - dailyMinutes
        - xp
        - level
        - streakCount
        - streakFreezes
        - weeklyXp
        - isPublicProfile
        - hideFromLeaderboard
        - anonymousMode
        - emailRemindersEnabled
        - theme
        - hasPaid
        - isGrandfathered
        - hasAccess
        - accessibleThroughDay
        - onboardingCompleted
        - createdAt
...
    OnboardingInput:
      type: object
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
          description: Map of quiz step id -> selected option value.
      required:
        - answers
    OnboardingProgram:
      type: object
      properties:
        programTitle:
          type: string
        programFocus:
          type: string
        workContext:
          type: ["string", "null"]
        highlights:
          type: array
          items:
            type: string
        estimatedWeeks:
          type: integer
      required:
        - programTitle
        - programFocus
        - highlights
        - estimatedWeeks
...
    QuizQuestionPublic:
      type: object
      properties:
        id:
          type: integer
        day:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        explanation:
          type: ["string", "null"]
          description: Shown after the learner answers, explaining why the correct option is right.
        orderIndex:
          type: integer
      required:
        - id
        - day
        - question
        - options
        - orderIndex
    CurriculumDayDetail:
      type: object
      properties:
        day:
          type: integer
        title:
          type: string
        theme:
          type: string
        lessonTitle:
          type: string
        lessonContent:
          type: string
        taskTitle:
          type: string
        taskDescription:
          type: string
        reflectionPrompt:
          type: string
        estimatedMinutes:
          type: integer
        xpReward:
          type: integer
        practicePrompt:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        audioDurationSec:
          type: ["integer", "null"]
        quiz:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionPublic"
      required:
        - day
        - title
        - theme
        - lessonTitle
        - lessonContent
        - taskTitle
        - taskDescription
        - reflectionPrompt
        - estimatedMinutes
        - xpReward
        - quiz
...
    DayAudio:
      type: object
      properties:
        audioUrl:
          type: string
        durationSec:
          type: ["integer", "null"]
      required:
        - audioUrl
        - durationSec
...
    PostAuthor:
      type: object
      properties:
        userId:
          type: integer
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        level:
          type: integer
        isPublicProfile:
          type: boolean
      required:
        - userId
        - level
        - isPublicProfile
    FeedPost:
      type: object
      properties:
        id:
          type: integer
        day:
          type: ["integer", "null"]
        kind:
          type: string
          enum: [win, reflection]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/PostAuthor"
        likeCount:
          type: integer
        likedByMe:
          type: boolean
        isMine:
          type: boolean
      required:
        - id
        - kind
        - content
        - createdAt
        - author
        - likeCount
        - likedByMe
        - isMine
    PostInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
        day:
          type: integer
        kind:
          type: string
          enum: [win, reflection]
      required:
        - content
...